Updated backpropagation functions from C to C++ to reduce memory errors.

+++ DeepNetwork.Infrastructure/Activation Functions/logistic_function.h
@@ -4,4 +4,4 @@
 
 void apply_logistic(matrix matrix);
 
-float calculate_logistic_derivative(float input);
+double calculate_logistic_derivative(double input);
